requirements.txt
# Core Framework
pandas==2.2.1
numpy==1.26.4

# Machine Learning
scikit-learn==1.4.1.post1
scipy>=1.12.0

# Visualization (keeping matplotlib for backward compatibility)
matplotlib>=3.8.3
seaborn==0.13.2
plotly==5.18.0
kaleido==0.2.1

# Network Analysis
networkx>=3.2.1
python-louvain==0.16

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
aiohttp==3.9.3
playwright>=1.42.0
pytest-playwright>=0.4.4

# Data Performance
fastparquet==2024.2.0
pyarrow>=15.0.0

# Configuration & Scheduling
python-dotenv==1.0.1
schedule==1.2.1
python-crontab==3.0.0

# Caching
diskcache==5.6.3

# Testing
pytest==8.0.0
pytest-cov==4.1.0
pytest-asyncio==0.23.5
tqdm
psutil
flask==3.0.0
flask-cors
cairosvg>=2.7,<3
